+++ src/main/java/fr/abes/lnevent/security/exception/DonneeIncoherenteBddException.java
@@ -1,0 +1,11 @@
+package fr.abes.lnevent.security.exception;
+
+public class DonneeIncoherenteBddException extends Exception{
+
+
+    private static final long serialVersionUID = 1L;
+
+    public DonneeIncoherenteBddException(String message){
+        super(message);
+    }
+}
